# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  role: UserProfileRole;
  /** @nullable */
  organizationId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
  industry?: string;
  currency?: string;
}

export interface OrganizationUpdate {
  /** @minLength 1 */
  name?: string;
  industry?: string;
  currency?: string;
}

export interface Organization {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  currency: string;
  createdAt: string;
  updatedAt?: string;
}

export interface FindingTypeStat {
  type: string;
  count: number;
  totalAtRisk: number;
}

export interface FindingSeverityStat {
  severity: string;
  count: number;
  totalAtRisk: number;
}

export interface OrganizationStats {
  /** Total estimated revenue at risk in organization currency */
  totalAtRisk: number;
  findingsCount: number;
  openFindingsCount: number;
  resolvedFindingsCount: number;
  recommendationsCount: number;
  pendingRecommendationsCount: number;
  byType: FindingTypeStat[];
  bySeverity: FindingSeverityStat[];
}

export type DataSourceInputType = typeof DataSourceInputType[keyof typeof DataSourceInputType];


export const DataSourceInputType = {
  csv: 'csv',
  hubspot: 'hubspot',
  pipedrive: 'pipedrive',
  google_sheets: 'google_sheets',
  stripe: 'stripe',
  asaas: 'asaas',
} as const;

export interface DataSourceInput {
  /** @minLength 1 */
  name: string;
  type: DataSourceInputType;
  description?: string;
}

export type DataSourceType = typeof DataSourceType[keyof typeof DataSourceType];


export const DataSourceType = {
  csv: 'csv',
  hubspot: 'hubspot',
  pipedrive: 'pipedrive',
  google_sheets: 'google_sheets',
  stripe: 'stripe',
  asaas: 'asaas',
} as const;

export type DataSourceStatus = typeof DataSourceStatus[keyof typeof DataSourceStatus];


export const DataSourceStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface DataSource {
  id: number;
  orgId: number;
  name: string;
  type: DataSourceType;
  status: DataSourceStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  rowCount?: number | null;
  /** @nullable */
  lastAnalyzedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface AnalysisResult {
  dataSourceId: number;
  findingsCreated: number;
  recommendationsCreated: number;
}

export type FindingType = typeof FindingType[keyof typeof FindingType];


export const FindingType = {
  overdue_invoice: 'overdue_invoice',
  inactive_customer: 'inactive_customer',
  stalled_opportunity: 'stalled_opportunity',
  contract_expiration: 'contract_expiration',
} as const;

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  open: 'open',
  acknowledged: 'acknowledged',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

/**
 * @nullable
 */
export type FindingMetadata = { [key: string]: unknown } | null;

export interface Finding {
  id: number;
  orgId: number;
  dataSourceId: number;
  type: FindingType;
  severity: FindingSeverity;
  title: string;
  description: string;
  /** Estimated revenue impact in organization currency */
  estimatedImpact: number;
  /**
     * Customer name, contract ID, or other entity identifier
     * @nullable
     */
  affectedEntity?: string | null;
  /** @nullable */
  daysOverdue?: number | null;
  status: FindingStatus;
  /** @nullable */
  metadata?: FindingMetadata;
  /** @nullable */
  resolvedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type FindingUpdateStatus = typeof FindingUpdateStatus[keyof typeof FindingUpdateStatus];


export const FindingUpdateStatus = {
  open: 'open',
  acknowledged: 'acknowledged',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

export interface FindingUpdate {
  status?: FindingUpdateStatus;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type RecommendationStatus = typeof RecommendationStatus[keyof typeof RecommendationStatus];


export const RecommendationStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  dismissed: 'dismissed',
} as const;

export interface Recommendation {
  id: number;
  orgId: number;
  findingId: number;
  priority: RecommendationPriority;
  title: string;
  description: string;
  /** Estimated recoverable revenue */
  estimatedRecovery: number;
  /**
     * Short action button label
     * @nullable
     */
  actionLabel?: string | null;
  status: RecommendationStatus;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type RecommendationUpdateStatus = typeof RecommendationUpdateStatus[keyof typeof RecommendationUpdateStatus];


export const RecommendationUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  dismissed: 'dismissed',
} as const;

export interface RecommendationUpdate {
  status?: RecommendationUpdateStatus;
}

export interface CsvUploadInput {
  file: Blob;
}