# lib/api-zod/src/generated/types/dataSourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARQON Revenue Engine V1 API
 * OpenAPI spec version: 0.1.0
 */
import type { DataSourceInputType } from './dataSourceInputType';

export interface DataSourceInput {
  /** @minLength 1 */
  name: string;
  type: DataSourceInputType;
  description?: string;
}